+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Value class now accepts utf8 string
 - Improves and unify plugins interfaces
 - Model plugins can now add and remove nodes and edges during the simulation
+- Model plugins: added beforeLoop() and afterLoop() methods
 - Increases unit-tests coverage of core components
 - Improves threads management
 - Improves TableWidget performance